refactor fxspot exceptions boundaries
