+++ backend/src/main/java/com/alligator/market/backend/instrument/asset/forex/fxspot/api/advice/FxSpotRestExceptionHandler.java
@@ -6,15 +6,15 @@ import com.alligator.market.backend.instrument.asset.forex.fxspot.api.command.cr
 import com.alligator.market.backend.instrument.asset.forex.fxspot.api.command.delete.controller.DeleteFxSpotController;
 import com.alligator.market.backend.instrument.asset.forex.fxspot.api.command.update.controller.UpdateFxSpotController;
 import com.alligator.market.backend.instrument.asset.forex.fxspot.api.query.list.controller.ListFxSpotsController;
+import com.alligator.market.backend.instrument.asset.forex.fxspot.application.exception.FxSpotAlreadyExistsException;
+import com.alligator.market.backend.instrument.asset.forex.fxspot.application.exception.FxSpotCreateException;
+import com.alligator.market.backend.instrument.asset.forex.fxspot.application.exception.FxSpotDeleteException;
 import com.alligator.market.backend.instrument.asset.forex.fxspot.application.exception.FxSpotInUseException;
+import com.alligator.market.backend.instrument.asset.forex.fxspot.application.exception.FxSpotNotFoundException;
+import com.alligator.market.backend.instrument.asset.forex.fxspot.application.exception.FxSpotUpdateException;
 import com.alligator.market.backend.instrument.asset.forex.reference.currency.application.exception.CurrencyNotFoundException;
 import com.alligator.market.domain.common.exception.DomainErrorCode;
-import com.alligator.market.domain.instrument.asset.forex.fxspot.exception.FxSpotAlreadyExistsException;
-import com.alligator.market.domain.instrument.asset.forex.fxspot.exception.FxSpotCreateException;
-import com.alligator.market.domain.instrument.asset.forex.fxspot.exception.FxSpotDeleteException;
-import com.alligator.market.domain.instrument.asset.forex.fxspot.exception.FxSpotNotFoundException;
 import com.alligator.market.domain.instrument.asset.forex.fxspot.exception.FxSpotSameCurrenciesException;
-import com.alligator.market.domain.instrument.asset.forex.fxspot.exception.FxSpotUpdateException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.Ordered;
 import org.springframework.core.annotation.Order;
