+++ backend/src/main/java/com/alligator/market/backend/instrument/asset/forex/fxspot/application/command/update/UpdateFxSpotService.java
@@ -1,8 +1,8 @@
 package com.alligator.market.backend.instrument.asset.forex.fxspot.application.command.update;
 
 import com.alligator.market.backend.instrument.asset.forex.fxspot.application.command.update.dto.UpdateFxSpotCommand;
+import com.alligator.market.backend.instrument.asset.forex.fxspot.application.exception.FxSpotNotFoundException;
 import com.alligator.market.domain.instrument.asset.forex.fxspot.FxSpot;
-import com.alligator.market.domain.instrument.asset.forex.fxspot.exception.FxSpotNotFoundException;
 import com.alligator.market.domain.instrument.asset.forex.fxspot.repository.FxSpotRepository;
 import lombok.extern.slf4j.Slf4j;
