+++ backend/src/main/java/com/alligator/market/backend/instrument/asset/forex/fxspot/persistence/jooq/repository/JooqFxSpotRepositoryAdapter.java
@@ -1,13 +1,13 @@
 package com.alligator.market.backend.instrument.asset.forex.fxspot.persistence.jooq.repository;
 
+import com.alligator.market.backend.instrument.asset.forex.fxspot.application.exception.FxSpotAlreadyExistsException;
+import com.alligator.market.backend.instrument.asset.forex.fxspot.application.exception.FxSpotCreateException;
+import com.alligator.market.backend.instrument.asset.forex.fxspot.application.exception.FxSpotDeleteException;
+import com.alligator.market.backend.instrument.asset.forex.fxspot.application.exception.FxSpotNotFoundException;
+import com.alligator.market.backend.instrument.asset.forex.fxspot.application.exception.FxSpotUpdateException;
 import com.alligator.market.backend.instrument.asset.forex.reference.currency.application.exception.CurrencyNotFoundException;
 import com.alligator.market.domain.instrument.asset.forex.fxspot.FxSpot;
 import com.alligator.market.domain.instrument.asset.forex.fxspot.classification.FxSpotTenor;
-import com.alligator.market.domain.instrument.asset.forex.fxspot.exception.FxSpotAlreadyExistsException;
-import com.alligator.market.domain.instrument.asset.forex.fxspot.exception.FxSpotCreateException;
-import com.alligator.market.domain.instrument.asset.forex.fxspot.exception.FxSpotDeleteException;
-import com.alligator.market.domain.instrument.asset.forex.fxspot.exception.FxSpotNotFoundException;
-import com.alligator.market.domain.instrument.asset.forex.fxspot.exception.FxSpotUpdateException;
 import com.alligator.market.domain.instrument.asset.forex.fxspot.repository.FxSpotRepository;
 import com.alligator.market.domain.instrument.asset.forex.reference.currency.Currency;
 import com.alligator.market.domain.instrument.asset.forex.reference.currency.vo.CurrencyCode;
